+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 1. [#93](https://github.com/influxdata/influxdb-client-java/issues/93): Add addTags and addFields helper functions to Point
 1. [#97](https://github.com/influxdata/influxdb-client-java/pull/97): Add the ability to specify the org and the bucket when creating the client
 
+### Bugs
+1. [#98](https://github.com/influxdata/influxdb-client-java/issues/98): @Column supports super class inheritance for write measurements
+
 ## 1.6.0 [2020-03-13]
 
 ### Features